refactor: re-use the pagination wrapper
