+++ spog/ui/crates/components/src/pagination.rs
@@ -1,0 +1,31 @@
+use patternfly_yew::prelude::*;
+use yew::prelude::*;
+
+#[derive(PartialEq, Properties)]
+pub struct PaginationWrappedProperties {
+    pub children: Children,
+    pub pagination: UsePagination,
+    pub total: Option<usize>,
+}
+
+#[function_component(PaginationWrapped)]
+pub fn pagination_wrapped(props: &PaginationWrappedProperties) -> Html {
+    html!(
+        <>
+            <div class="pf-v5-u-p-sm">
+                <SimplePagination
+                    pagination={props.pagination.clone()}
+                    total={props.total}
+                />
+            </div>
+            { for props.children.iter() }
+            <div class="pf-v5-u-p-sm">
+                <SimplePagination
+                    pagination={props.pagination.clone()}
+                    total={props.total}
+                    position={PaginationPosition::Bottom}
+                />
+            </div>
+        </>
+    )
+}
